Update to regex pattern matching

+++ src/vaporize/index.ts
@@ -5,6 +5,11 @@ import fs from "node:fs/promises";
 import { randomUUID } from "node:crypto";
 import path from "node:path";
 
+interface FileData {
+    ext: string;
+    file: string;
+}
+
 /**
  * The error list.
  * @type {Array}
@@ -34,6 +39,40 @@ const getFileData = async (filePath: string): Promise<any> => {
     throw new Error("Invalid file type.");
 }
 
+const removeUnusedDependencies = (fileData: FileData, dependencies: Array<string>, isEsm: boolean): boolean => {
+    const codeWithoutWhiteSpace: string = fileData.file.replace(/\s/g, "");
+    const dependencyStatements = isEsm
+      ? lib.getImports(dependencies, codeWithoutWhiteSpace)
+      : lib.getRequirements(dependencies, codeWithoutWhiteSpace);
+    const variableNames = lib.getVariableNames(dependencyStatements, dependencies, isEsm);
+    const unusedReferences: Array<string> = [];
+    for (let i = 0; i < variableNames.length; i++) {
+        lib.findVariableReferences(variableNames[i], codeWithoutWhiteSpace, unusedReferences);
+    }
+
+    if (unusedReferences.length === 0) {
+        console.log("No unused dependencies found.");
+        return false;
+    }
+
+    for (let i = 0; i < unusedReferences.length; i++) {
+        const pattern = isEsm
+        ? String.raw`import[/\s/]*${unusedReferences[i]}[\/\s\/]*from[\/\s\/]*["'][A-Za-z0-9\-\/\.\:]*["'][\/\s\/\;]*`
+        : String.raw`(const|let|var)[\/\s\/]*${unusedReferences[i]}[\/\s\/]*=[\/\s\/]*require\(["'][A-Za-z0-9\-\/\.\:]*["']\)[\/\s\/\;]*`;
+        fileData.file = fileData.file.replace(new RegExp(pattern, "gm"), "");
+    }
+
+    return true;
+}
+
+const writeTempFile = async (filePath: string, fileData: FileData): Promise<string> => {
+    const temp = path.resolve(filePath);
+    const targetDirectory = path.resolve(temp.replace(path.basename(temp), "").replace("/src", ""), path.dirname(filePath));
+    const tempFilePath = path.join(targetDirectory, randomUUID() + fileData.ext);
+    await fs.writeFile(tempFilePath, fileData.file);
+    return tempFilePath;
+}
+
 /**
  * Removes references to unused dependencies in a JavaScript or TypeScript file.
  * @param {string} filePath The path to the file.
@@ -41,68 +80,30 @@ const getFileData = async (filePath: string): Promise<any> => {
  */
 export const vaporize = async (filePath: string) => {
     const fileData = await getFileData(filePath);
-
     const dependencies: Array<string> = precinct(fileData.file);
     if (dependencies.length === 0) {
         return;
     }
 
-    let dependencyStatements: Array<string> = [];
-    let variableNames: Array<string> = [];
-    const unusedReferences: Array<string> = [];
-    const codeWithoutWhiteSpace: string = fileData.file.replace(/\s/g, "");
+    let hasRemovedUnusedDependencies: boolean;
     if ((fileData.ext === EXTENSION.js && fileData.file.includes("import")) || fileData.ext === EXTENSION.ts) {
-        dependencyStatements = lib.getImports(dependencies, codeWithoutWhiteSpace);
-        variableNames = lib.getVariableNames(dependencyStatements, dependencies, true);
-        for (let i = 0; i < variableNames.length; i++) {
-            const variableName = variableNames[i];
-            lib.findVariableReferences(variableName, codeWithoutWhiteSpace, unusedReferences);
-        }
-
-        if (unusedReferences.length === 0) {
-            console.log("No unused dependencies found.");
-            return;
-        }
-
-        // iterate over each unused reference & replace with empty string
-        for (let i = 0; i < unusedReferences.length; i++) {
-            fileData.file = fileData.file.replace(new RegExp(String.raw`import[/\s/]*${unusedReferences[i]}[/\s/]*from[/\s/]*["'][A-Za-z0-9\-\:]*["'][/\s/\;]*`, "gm"), "");
-        }
+        hasRemovedUnusedDependencies = removeUnusedDependencies(fileData, dependencies, true);
     } else if (fileData.ext === EXTENSION.cjs || fileData.ext === EXTENSION.js) {
-        dependencyStatements = lib.getRequirements(dependencies, codeWithoutWhiteSpace);
-        variableNames = lib.getVariableNames(dependencyStatements, dependencies, false);
-        for (let i = 0; i < variableNames.length; i++) {
-            const variableName = variableNames[i];
-            lib.findVariableReferences(variableName, codeWithoutWhiteSpace, unusedReferences);
-        }
-
-        if (unusedReferences.length === 0) {
-            console.log("No unused dependencies found.");
-            return;
-        }
-        
-        // iterate over each unused reference & replace with empty string
-        for (let i = 0; i < unusedReferences.length; i++) {
-            fileData.file = fileData.file.replace(new RegExp(String.raw`(const|let|var)[/\s/]*${unusedReferences[i]}[/\s/]*=[/\s/]*require\(["'][A-Za-z0-9\-]*["']\)[/\s/\;]*`, "gm"), "");
-        }
+        hasRemovedUnusedDependencies = removeUnusedDependencies(fileData, dependencies, false);
     }
 
-    // This will only work for files within the vaporize directory.
-    // I need to figure out how I can read & write files from outside of the Vaporize project
-    const temp = path.resolve(filePath);
-    const targetDirectory = path.resolve(temp.replace(path.basename(temp), "").replace("/src", ""), path.dirname(filePath));
-    const tempFilePath = path.join(targetDirectory, randomUUID() + fileData.ext);
-    await fs.writeFile(tempFilePath, fileData.file);
-    const readTempFileResult = await lib.executeFilePromise(tempFilePath);
-    if (typeof (readTempFileResult) !== "boolean" && typeof (readTempFileResult) !== "undefined") {
-        ERROR_LIST.push(readTempFileResult);
-    }
-
-    // delete the file
-    await fs.unlink(tempFilePath);
-
-    // replace the code in the file with the sanitized code if executing the code did not throw any errors
-    if (ERROR_LIST.length === 0) {
-        await fs.writeFile(filePath, fileData.file);
+    if (hasRemovedUnusedDependencies) {
+        const tempFilePath = await writeTempFile(filePath, fileData);
+        const readTempFileResult = await lib.executeFilePromise(tempFilePath);
+        if (typeof (readTempFileResult) !== "boolean" && typeof (readTempFileResult) !== "undefined") {
+            ERROR_LIST.push(readTempFileResult);
+        }
+    
+        // delete the file
+        await fs.unlink(tempFilePath);
+    
+        if (ERROR_LIST.length === 0) {
+            await fs.writeFile(filePath, fileData.file);
+        }
     }
 }
